+++ CHANGELOG.md
@@ -3,7 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
-## [v1.0-1.16.5] - 2021-11-02
+## [v1.0.1-1.16.5] - 2021-11-19
+### Changed
+- The vanilla effect renderer now supports multiple columns
+- Changed default values for a few config options
+
+## [v1.0.0-1.16.5] - 2021-11-02
 - Initial release
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/